test(reservation): api 통합 테스트와 controller 슬라이스 테스트 분리
