+++ src/test/java/roomescape/common/TestArgumentResolverConfig.java
@@ -1,0 +1,15 @@
+package roomescape.common;
+
+import java.util.List;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@TestConfiguration
+public class TestArgumentResolverConfig implements WebMvcConfigurer {
+
+    @Override
+    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
+        resolvers.add(new TestLoginMemberArgumentResolver());
+    }
+}
